Use a dictionary instead of a switch
Interesting that a dictionary does not accept a Character for a key,
only a String.  This means that romanNumeral which is a Character should
be casted as a String - Swift complains otherwise.

I'm also starting to appreciate the safety in Swift, wherein since it is
possible that a value would be nil from our array, we should wrap it in
this syntax.